requirements.txt
groq
pandas
pytest
python-dotenv
streamlit
tabulate
